explicit cast (removes warnings)

+++ compiler/generator/occurences.cpp
@@ -40,7 +40,7 @@ static int xVariability(int v, int r)
     // faustassert(v < 3);				// kKonst=0, kBlock=1, kSamp=2
     // faustassert(r==0 | v==2);
     if (r > 1) r = 1;
-    return min(3, v + r);
+    return (int)min(3, v + r);
 }
 
 //-------------------------------------------------
